2013-10-04 15:58 UTC+0200 Viktor Szakáts (vszakats users.noreply.github.com)   * config/win/msvc.mk   * contrib/hbpost.hbm     + added code to enable -sdl option. not enabled, because some       3rd party components fail to compile
  * contrib/hbwin/hbolesrv.c
    ! fixed char -> TCHAR after recent unicodification

  * contrib/hbtip/utils.c
  * src/common/hbdate.c
  * src/rtl/gtwvt/gtwvt.c
    * added missing casts shown by msvc

  * src/3rd/png/Makefile
  * src/3rd/png/png.dif
  * src/3rd/png/pnglconf.h
    * disabled STDIO dependent parts for all platforms
    * disabled PNG_CONVERT_tIME_SUPPORTED to avoid using
      unsafe RTL calls (as shown by msvc)

  * README.md
    * wording

+++ contrib/hbtip/utils.c
@@ -84,7 +84,7 @@ HB_FUNC( TIP_TIMESTAMP )
 
       /* For compatibility, seconds() value */
       if( HB_ISNUM( 2 ) )
-         hb_timeDecode( hb_parnd( 2 ) * 1000,
+         hb_timeDecode( ( long ) ( hb_parnd( 2 ) * 1000 ),
                         &iHour, &iMinute, &iSecond, &iMSec );
       else
          iHour = iMinute = iSecond = 0;
